# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AImanju Backend API",
    "version": "0.1.0",
    "description": "当前仓库已实现路由；标记待接入的路由始终返回 503。服务供应商和前端宽松字段仍需联调。"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/api/login/password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "密码登录",
        "security": [],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "const": 200
                    },
                    "message": {
                      "type": "string",
                      "const": "ok"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AuthTokens"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ],
                  "additionalProperties": false
                },
                "example": {
                  "code": 200,
                  "message": "ok",
                  "data": {
                    "access_token": "<jwt>",
                    "refresh_token": "<opaque>"
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求参数无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "凭证无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在或不属于当前账号",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本或幂等冲突；画布冲突包含 data.current_version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "依赖不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordLogin"
              }
            }
          }
        }
      }
    },
    "/api/sms/send": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "发送短信验证码（仅显式开发模拟）",
        "security": [],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "const": 200
                    },
                    "message": {
                      "type": "string",
                      "const": "ok"
                    },
                    "data": {
                      "$ref": "#/components/schemas/SmsSendResult"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ],
                  "additionalProperties": false
                },
                "example": {
                  "code": 200,
                  "message": "ok",
                  "data": {
                    "mock": true
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求参数无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "凭证无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在或不属于当前账号",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本或幂等冲突；画布冲突包含 data.current_version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "发送过于频繁或达到每日上限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "依赖不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SmsSend"
              }
            }
          }
        }
      }
    },
    "/api/login/sms": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "短信登录；首次使用手机号自动注册个人账号",
        "security": [],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "const": 200
                    },
                    "message": {
                      "type": "string",
                      "const": "ok"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AuthTokens"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ],
                  "additionalProperties": false
                },
                "example": {
                  "code": 200,
                  "message": "ok",
                  "data": {
                    "access_token": "<jwt>",
                    "refresh_token": "<opaque>"
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求参数无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "凭证无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在或不属于当前账号",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本或幂等冲突；画布冲突包含 data.current_version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "依赖不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SmsLogin"
              }
            }
          }
        }
      }
    },
    "/api/login/wechat": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "微信登录（待接入）",
        "security": [],
        "responses": {
          "400": {
            "description": "请求参数无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "外部服务未配置或功能待接入",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WechatLogin"
              }
            }
          }
        }
      }
    },
    "/api/user/bindPhone": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "绑定手机号（仅显式开发模拟）",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope"
                },
                "example": {
                  "code": 200,
                  "message": "ok",
                  "data": {}
                }
              }
            }
          },
          "400": {
            "description": "请求参数无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "凭证无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在或不属于当前账号",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本或幂等冲突；画布冲突包含 data.current_version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "依赖不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SmsLogin"
              }
            }
          }
        }
      }
    },
    "/api/auth/refresh": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "使用 Bearer refresh token 轮换会话",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "const": 200
                    },
                    "message": {
                      "type": "string",
                      "const": "ok"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AuthTokens"
                    }
                  },
                  "required": [
                    "code",
                    "message",
                    "data"
                  ],
                  "additionalProperties": false
                },
                "example": {
                  "code": 200,
                  "message": "ok",
                  "data": {
                    "access_token": "<jwt>",
                    "refresh_token": "<opaque>"
                  }
                }
              }
            }
          },
          "400": {
            "description": "请求参数无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "凭证无效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "权限不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "资源不存在或不属于当前账号",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "版本或幂等冲突；画布冲突包含 data.current_version",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "依赖不可用",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Envelope": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string",
            "example": "ok"
          },
          "data": {}
        },
        "required": [
          "code",
          "message",
          "data"
        ],
        "additionalProperties": false
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "nullable": true
          }
        },
        "required": [
          "code",
          "message"
        ],
        "additionalProperties": false
      },
      "PasswordLogin": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "additionalProperties": false
      },
      "AuthTokens": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string"
          },
          "refresh_token": {
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "refresh_token"
        ],
        "additionalProperties": false
      },
      "SmsSend": {
        "type": "object",
        "properties": {
          "mobile": {
            "type": "string"
          },
          "sendType": {
            "type": "string",
            "enum": [
              "login_register",
              "bind_phone"
            ]
          }
        },
        "required": [
          "mobile",
          "sendType"
        ],
        "additionalProperties": false
      },
      "SmsLogin": {
        "type": "object",
        "properties": {
          "mobile": {
            "type": "string"
          },
          "captcha": {
            "type": "string"
          }
        },
        "required": [
          "mobile",
          "captcha"
        ],
        "additionalProperties": false
      },
      "SmsSendResult": {
        "type": "object",
        "properties": {
          "mock": {
            "type": "boolean",
            "const": true
          }
        },
        "required": [
          "mock"
        ],
        "additionalProperties": false
      },
      "WechatLogin": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          }
        },
        "required": [
          "code"
        ],
        "additionalProperties": false
      },
...
    }
  }
}